Revision
This is the revision stage of the writing process. I did not use the Elements of Style, because the online version that we were told we could use did not contain the pages we needed for this stage. I tried to get the book from the library but could not find it, and I did not have time to go to the bookstore at that time.